requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
attrs==25.3.0
base58==2.1.1
bitarray==3.4.2
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
cytoolz==1.0.1
eth_abi==5.2.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
Flask==3.1.1
frozenlist==1.7.0
hexbytes==1.3.1
idna==3.10
ipfshttpclient==0.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multiaddr==0.0.9
multidict==6.4.4
netaddr==1.3.0
numpy==2.3.0
packaging==25.0
parsimonious==0.10.0
pefile==2023.2.7
pillow==11.2.1
propcache==0.3.2
py-solc-x==2.0.4
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyunormalize==16.0.0
pywin32==310
pywin32-ctypes==0.2.3
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
setuptools==80.9.0
six==1.17.0
stepic==0.5.0
toolz==1.0.0
types-requests==2.32.4.20250611
typing_extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
varint==1.0.2
web3==7.12.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
